requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
